+++ database/migrations/2021_11_26_040142_alter_stations_table.php
@@ -13,16 +13,18 @@ class AlterStationsTable extends Migration
      */
     public function up()
     {
-        Schema::table('stations1', function (Blueprint $table) {
-            $table->dropColumn([
-                'region_code', 'tran_company_code', 'station_code',
-                'tran_company_full_name', 'tran_company_short_name',
-                'old_name', 'number_change', 'value_change',
-                'parent_flag'
-            ]);
-            $table->integer('transport_id');
-            $table->dropUnique('stations_name_unique');
-        });
+        if (Schema::hasTable('stations')) {
+            Schema::table('stations', function (Blueprint $table) {
+                $table->dropColumn([
+                    'region_code', 'tran_company_code', 'station_code',
+                    'tran_company_full_name', 'tran_company_short_name',
+                    'old_name', 'number_change', 'value_change',
+                    'parent_flag'
+                ]);
+                $table->integer('transport_id');
+                $table->dropUnique('stations_name_unique');
+            });
+        }
     }
 
     /**
@@ -32,17 +34,19 @@ class AlterStationsTable extends Migration
      */
     public function down()
     {
-        Schema::table('stations1', function (Blueprint $table) {
-            $table->integer('region_code')->nullable()->after('id');
-            $table->integer('tran_company_code')->nullable()->after('region_code');
-            $table->integer('station_code')->nullable()->after('tran_company_code');
-            $table->string('tran_company_full_name')->nullable()->after('station_code');
-            $table->string('tran_company_short_name')->nullable()->after('tran_company_full_name');
-            $table->string('old_name')->nullable()->after('tran_company_short_name');
-            $table->integer('number_change')->nullable()->after('name');
-            $table->integer('value_change')->nullable()->after('number_change');
-            $table->boolean('parent_flag')->default(true)->after('count_estates');
-            $table->dropColumn('transport_id');
-        });
+        if (Schema::hasTable('stations')) {
+            Schema::table('stations', function (Blueprint $table) {
+                $table->integer('region_code')->nullable()->after('id');
+                $table->integer('tran_company_code')->nullable()->after('region_code');
+                $table->integer('station_code')->nullable()->after('tran_company_code');
+                $table->string('tran_company_full_name')->nullable()->after('station_code');
+                $table->string('tran_company_short_name')->nullable()->after('tran_company_full_name');
+                $table->string('old_name')->nullable()->after('tran_company_short_name');
+                $table->integer('number_change')->nullable()->after('name');
+                $table->integer('value_change')->nullable()->after('number_change');
+                $table->boolean('parent_flag')->default(true)->after('count_estates');
+                $table->dropColumn('transport_id');
+            });
+        }
     }
 }
